+++ internal/service/projectipaccesslist/model_project_ip_access_list_test.go
@@ -1,7 +1,6 @@
 package projectipaccesslist_test
 
 import (
-	"context"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-framework-timeouts/resource/timeouts"
@@ -207,7 +206,7 @@ func TestNewTfProjectIPAccessListDSModel(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			resultModel, _ := projectipaccesslist.NewTfProjectIPAccessListDSModel(context.Background(), tc.sdkModel)
+			resultModel, _ := projectipaccesslist.NewTfProjectIPAccessListDSModel(t.Context(), tc.sdkModel)
 
 			assert.Equal(t, tc.expectedResult, resultModel)
 		})