+++ internal/service/searchdeployment/model_test.go
@@ -1,14 +1,15 @@
 package searchdeployment_test
 
 import (
-	"context"
 	"reflect"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/searchdeployment"
-	"go.mongodb.org/atlas-sdk/v20250219001/admin"
+
+	// TODO: update before merging to master: "go.mongodb.org/atlas-sdk/v20250219001/admin"
+	"github.com/mongodb/atlas-sdk-go/admin"
 )
 
 type sdkToTFModelTestCase struct {
@@ -25,6 +26,7 @@ const (
 	clusterName       = "Cluster0"
 	instanceSize      = "S20_HIGHCPU_NVME"
 	nodeCount         = 2
+	earProvider       = "AWS"
 )
 
 func TestSearchDeploymentSDKToTFModel(t *testing.T) {
@@ -42,20 +44,22 @@ func TestSearchDeploymentSDKToTFModel(t *testing.T) {
 						NodeCount:    nodeCount,
 					},
 				},
+				EncryptionAtRestProvider: admin.PtrString(earProvider),
 			},
 			expectedTFModel: &searchdeployment.TFSearchDeploymentRSModel{
-				ID:          types.StringValue(dummyDeploymentID),
-				ClusterName: types.StringValue(clusterName),
-				ProjectID:   types.StringValue(dummyProjectID),
-				StateName:   types.StringValue(stateName),
-				Specs:       tfSpecsList(t, instanceSize, nodeCount),
+				ID:                       types.StringValue(dummyDeploymentID),
+				ClusterName:              types.StringValue(clusterName),
+				ProjectID:                types.StringValue(dummyProjectID),
+				StateName:                types.StringValue(stateName),
+				Specs:                    tfSpecsList(t, instanceSize, nodeCount),
+				EncryptionAtRestProvider: types.StringValue(earProvider),
 			},
 		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			resultModel, diags := searchdeployment.NewTFSearchDeployment(context.Background(), tc.clusterName, tc.SDKResp, nil, false)
+			resultModel, diags := searchdeployment.NewTFSearchDeployment(t.Context(), tc.clusterName, tc.SDKResp, nil, false)
 			if diags.HasError() {
 				t.Errorf("unexpected errors found: %s", diags.Errors()[0].Summary())
 			}
@@ -94,7 +98,7 @@ func TestSearchDeploymentTFModelToSDK(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			apiReqResult := searchdeployment.NewSearchDeploymentReq(context.Background(), tc.tfModel)
+			apiReqResult := searchdeployment.NewSearchDeploymentReq(t.Context(), tc.tfModel)
 			if !reflect.DeepEqual(apiReqResult, tc.expectedSDKReq) {
 				t.Errorf("created sdk model did not match expected output")
 			}
@@ -104,7 +108,7 @@ func TestSearchDeploymentTFModelToSDK(t *testing.T) {
 
 func tfSpecsList(t *testing.T, instanceSize string, nodeCount int64) basetypes.ListValue {
 	t.Helper()
-	tfSpecsList, diags := types.ListValueFrom(context.Background(), searchdeployment.SpecObjectType, []searchdeployment.TFSearchNodeSpecModel{
+	tfSpecsList, diags := types.ListValueFrom(t.Context(), searchdeployment.SpecObjectType, []searchdeployment.TFSearchNodeSpecModel{
 		{
 			InstanceSize: types.StringValue(instanceSize),
 			NodeCount:    types.Int64Value(nodeCount),